+++ frontend/package.json
@@ -29,6 +29,7 @@
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
     "autoprefixer": "10.4.17",
+    "eslint": "^8.56.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-prettier": "^9.1.0",
@@ -37,17 +38,16 @@
     "eslint-plugin-simple-import-sort": "^12.0.0",
     "eslint-plugin-svelte": "^2.35.1",
     "eslint-plugin-unicorn": "^51.0.1",
-    "eslint": "^8.56.0",
     "postcss": "^8.4.35",
-    "prettier-plugin-svelte": "^3.2.1",
     "prettier": "^3.2.5",
-    "svelte-check": "^3.6.4",
+    "prettier-plugin-svelte": "^3.2.1",
     "svelte": "^4.2.11",
+    "svelte-check": "^3.6.4",
     "tailwindcss": "^3.4.1",
     "tslib": "^2.6.2",
     "typescript": "^5.3.3",
-    "vite-plugin-tailwind-purgecss": "0.2.0",
-    "vite": "^5.1.3"
+    "vite": "^5.1.3",
+    "vite-plugin-tailwind-purgecss": "0.2.0"
   },
   "type": "module",
   "dependencies": {
@@ -62,6 +62,8 @@
     "keycloak-js": "^23.0.6",
     "mermaid": "^10.8.0",
     "minio": "^7.1.3",
-    "plotly.js-dist": "^2.29.1"
+    "plotly.js-dist": "^2.29.1",
+    "svelte-feather-icons": "^4.1.0",
+    "svelte-plotly.js": "^0.3.2"
   }
 }
