+++ frontend/src/modals/rename_project_modal.svelte
