+++ frontend/src/modals/submit_new_project_modal.svelte
