+++ frontend/src/modals/submit_new_secret_modal.svelte
