+++ frontend/src/modals/text_input_modal.svelte
