+++ frontend/src/routes/projects/dryruns/[dry_run]/symbol-for-action.svelte
@@ -13,6 +13,7 @@
 	import { pausedDryRuns } from '../../../../stores/stores.js';
 	import { requestGraphQLClient } from '$lib/graphqlUtils.js';
 	import { displayAlert } from '../../../../utils/alerts_utils.js';
+	import { getModalStore } from '@skeletonlabs/skeleton';
 
 	const modalStore = getModalStore();
