+++ frontend/src/utils/alerts_utils.ts
@@ -1,7 +1,12 @@
 // import type { ModalSettings } from '@skeletonlabs/skeleton';
-import { modalStore, type ModalSettings } from '@skeletonlabs/skeleton';
+import { getModalStore } from '@skeletonlabs/skeleton';
+import type { ModalSettings } from '@skeletonlabs/skeleton';
+
 
 export const displayAlert = (title: string, body: string, timeout = 2000) => {
+	
+	const modalStore = getModalStore();
+
 	const alertModal: ModalSettings = {
 		type: 'alert',
 		title,
